divide and rule
