+++ public/index.html
@@ -34,8 +34,8 @@
       <!-- Количество фильмов -->
     </section>
   </footer>
-  <!-- <script type="module" src="./bundle.js"></script> -->
-  <script type="module" src="../src/main.js"></script>
+  <script type="module" src="./bundle.js"></script>
+  <!-- <script type="module" src="../src/main.js"></script> -->
 </body>
 
 </html>
